add another copy of MIT license for clarity
All the content in this repository is open source and licensed under the MIT license. Add an additional copy of the MIT license into the root pom.xml file so you can't miss it. For brevity and to avoid red tape, we do not duplicate the license text on every darn source code file, but it is of course all MIT licensed.